C4

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@
 
 void print_number(int n)
 {
-long int num = n;
+unsigned int num = n;
 
 if (n < 0)
 {
